requirements.txt
pandas==2.2.2
openpyxl==3.1.5
google-cloud-bigquery==3.25.0
pyarrow==17.0.0
apache_airflow==2.10.1